# verseFactory_back/src/main/resources/openapi/openapi-admin-v1.yaml
openapi: 3.0.3
info:
  title: Template Admin API
  version: 1.0.0
  description: API pour les opérations d'administration (authentifiées)
servers:
  - url: /api/v1/admin
    description: Serveur local
security:
  - OAuth: []
paths:
  /pets:
    get:
      tags:
        - Pet
      summary: Get all pets (Admin only)
      operationId: getAllPets
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PetDto'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /pets/{id}:
    get:
      tags:
        - Pet
      summary: Get a pet by ID (Admin only)
      operationId: getPetById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PetDto'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /factory:
    get:
      tags:
        - Factory
      summary: Get factory details of the connected user
      operationId: getConnectedUserFactory
      security:
        - OAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FactoryDto'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
components:
  securitySchemes:
    OAuth:
      flows:
        authorizationCode:
          authorizationUrl: '{{authUrl}}'
          tokenUrl: '{{accessTokenUrl}}'
          refreshUrl: '{{callbackUrl}}'
          scopes:
            openid: ''
      type: oauth2
  schemas:
    FactoryDto:
      type: object
      required:
        - id
        - userId
        - balance
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        balance:
          type: number
        lastUpdatedAt:
          type: string
          format: date-time
    PetDto:
      type: object
      required:
        - id
        - name
        - rarity
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        rarity:
          type: string
        incomePerSecond:
          type: number
        baseCost:
          type: number
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
  responses:
...
    UnauthorizedError:
      description: Non authentifié
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ForbiddenError:
      description: Non autorisé
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundError:
      description: Ressource introuvable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'